pyproject.toml
[project]
name = "flux2-generator"
version = "0.1.0"
description = "Flux2 Image Generation for Apple Silicon Mac"
readme = "README.md"
requires-python = ">=3.11,<3.14"
license = { text = "MIT" }

dependencies = [
    # PyTorch (MPS support for Apple Silicon)
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "torchaudio>=2.1.0",

    # Hugging Face ecosystem
    "diffusers>=0.30.0",
    "transformers>=4.40.0",
    "accelerate>=0.30.0",
    "huggingface-hub>=0.23.0",

    # Tokenizer dependencies
    "sentencepiece>=0.2.0",
    "protobuf>=4.25.0",

    # Image processing
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.scripts]
generate = "generate:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]

[tool.mypy]
python_version = "3.12"
strict = true
